* Полностью реализована система режимов

+++ Assets/Scripts/SceneLoader.cs
@@ -8,18 +8,18 @@ public class SceneLoader : MonoBehaviour
     private static SceneLoader _insance;
     private GameModeType _gameModeType;
     private int _players;
-    private float _time;
+    private int _time;
 
     public static SceneLoader instance => _insance;
     public GameModeType gameModeType => _gameModeType;
     public int players => _players;
-    public float time => _time;
+    public int time => _time;
 
     public void StartGame(GameModeType gm, int players, int time)
     {
         _gameModeType = gm;
         _players = players;
-        _time = (int)time;
+        _time = time;
         SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
     }
